requirements.txt

biopython
dgllife
h5py
igraph
matplotlib
networkx
numpy
pandas
pandas-flavor
ray[data,train,tune,serve]
requests
scikit-learn
scipy
seaborn
tensorboard
tqdm
urllib3
torch
torchvision
torchaudio
gensim
lifelines
wandb
async_timeout
torch_geometric
deepsmiles
beautifulsoup4